requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.4.26
charset-normalizer==3.4.1
distro==1.9.0
flake8==7.2.0
groq==0.23.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mccabe==0.7.0
pycodestyle==2.13.0
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
